Add hosts to DB.  Add views for Running, Finished and Ready jobs

+++ ServerLessDb/Jobs/JobStart.sql
@@ -1,7 +1,7 @@
 ﻿CREATE TABLE [dbo].[JobStart]
 (
 	[JobId] UNIQUEIDENTIFIER NOT NULL, 	
-	[Attempt] SMALLINT NOT NULL, 
+	[Attempt] INT NOT NULL, 
     [StartAt] DATETIMEOFFSET NOT NULL, 
     [HostName] NVARCHAR(50) NOT NULL, 
     CONSTRAINT [PK_JobAttempt] PRIMARY KEY ([JobId], [Attempt]),
